+++ packages/system-access/src/base/prompts.ts
@@ -1,0 +1,62 @@
+import { existsSync } from 'fs';
+import type { PromptObject } from 'prompts';
+
+const username: PromptObject<string> = {
+    type: 'text',
+    name: 'username',
+    message: 'Username:'
+};
+
+const password: PromptObject<string> = {
+    type: 'password',
+    name: 'password',
+    message: 'Password:'
+};
+
+const serviceKeysPath: PromptObject<string> = {
+    type: 'text',
+    name: 'path',
+    message: 'Please provide the service keys as file:',
+    validate: (input) => existsSync(input)
+};
+
+const storeCredentials: PromptObject<string> = {
+    type: 'confirm',
+    name: 'store',
+    message: 'Do you want to store your credentials in the secure storage?',
+    initial: true
+};
+
+const systemName: PromptObject<string> = {
+    type: 'text',
+    name: 'name',
+    message: 'System name:',
+    validate: (input) => !!input
+};
+
+/**
+ * Export map of questions for usage with the prompts modules
+ */
+export const questions = {
+    username,
+    password,
+    serviceKeysPath,
+    storeCredentials,
+    systemName
+} as const;
+type keys = keyof typeof questions;
+
+/**
+ * Generate a map of questions for usages with inquirer (e.g. in yeoman)
+ */
+export const inquirer: { [key in keys]: object } = {} as { [key in keys]: object };
+for (const key in questions) {
+    const question = questions[key as keys];
+    inquirer[key as keys] = {
+        type: question.type === 'text' ? 'input' : question.type,
+        name: question.name,
+        message: question.message,
+        validate: question.validate,
+        default: question.initial
+    };
+}